ttempt at moving code within the new code driven structure without changing the logic
