+++ examples/energy-management-app/linux/BUILD.gn
@@ -145,6 +145,7 @@ source_set("test-evse-targets-store") {
 
   deps = [
     "${chip_root}/examples/energy-management-app/energy-management-common",
+    "${chip_root}/src/app/clusters/electrical-energy-measurement-server:electrical-energy-measurement-server",
     "${chip_root}/src/lib",
   ]
 }
