+++ src/app/clusters/electrical-energy-measurement-server/app_config_dependent_sources.gni
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "ElectricalEnergyMeasurementCluster.cpp",
+  "CodegenIntegration.cpp",
+  "CodegenIntegration.h",
   "EnergyReportingTestEventTriggerHandler.h",
   "electrical-energy-measurement-server.h",
 ]
